Refactor `audio` package
